requirements.txt
django
djangorestframework
neat-python
gymnasium[classic-control]
numpy
django-allauth
requests
